# docs/edgelet/edgelet-api-v1-openapi.yaml
openapi: 3.1.0
info:
  title: Edgelet API
  version: 1.0.3
  description: >
    Edgelet API v1 baseline for Edgelet daemon and CLI.
    Canonical namespace is /v1. Any path/schema/auth deviation must be
    explicitly approved before implementation.
servers:
  - url: https://edgelet.default.svc.bridge.local
    description: HTTPS/WSS transport for microservices and local clients
  - url: http+unix://%2Frun%2Fedgelet%2Fedgelet.sock
    description: Unix socket transport for CLI/admin operations
security:
  - BearerAuth: []
paths:
...
  /v1/models:
    get:
      tags: [Models]
      summary: List local models
      responses:
        "200":
          description: Model list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/models:pull:
    post:
      tags: [Models]
      summary: Pull model artifacts (async)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelPullRequest"
      responses:
        "202":
          description: Pull accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPullOperationStatusResponse"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/models:prune:
    post:
      tags: [Models]
      summary: Prune dangling models
      parameters:
        - in: query
          name: mode
          required: false
          schema:
            type: string
            enum: [dangling]
            default: dangling
      responses:
        "200":
          description: Prune result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPruneResponse"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    InvalidArgument:
      description: Invalid argument
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
    ApiError:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            requestId:
              type: string
...
    ModelListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ModelItem"
            count:
              type: integer
...
    ModelItem:
      type: object
      properties:
        name:
          type: string
        source:
          type: string
          enum: [local, managed]
          description: Provenance on this node. List and inspect both include source.
        uuid:
          type: string
          description: Controller model uuid when source is managed (inspect only).
        bindRefCount:
          type: integer
          description: Microservices that currently bind this name (inspect only).
        repo:
          type: string
        revision:
          type: string
        registryId:
          type: integer
        files:
          type: array
          items:
            type: string
        format:
          type: string
        state:
          type: string
          enum: [Pending, Pulling, Ready, Failed]
        generation:
          type: integer
          format: int64
        observedGeneration:
          type: integer
          format: int64
        revisionFloating:
          type: boolean
        totalBytes:
          type: integer
          format: int64
        lastError:
          type: string
        resolvedRevision:
          type: string
        digest:
          type: string
    ModelPullRequest:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name:
          type: string
        repo:
          type: string
          description: Repository path without host. Required with registryId when upserting.
        revision:
          type: string
        registryId:
          type: integer
          description: Local registry id. Required with repo when upserting.
        files:
          type: array
          items:
            type: string
          description: HF file paths or globs. Ignored for OCI.
        format:
          type: string
          description: Hint only (gguf, safetensors, onnx, pytorch, tensorrt, unknown).
    ModelPullOperationStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            operationId:
              type: string
            status:
              type: string
              enum: [running, succeeded, failed]
            progress:
              type: integer
            name:
              type: string
            bytesDownloaded:
              type: integer
              format: int64
            bytesTotal:
              type: integer
              format: int64
            registryId:
              type: integer
            error:
              type: string
            startedAt:
              type: string
            endedAt:
              type: string
    ModelPruneResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            mode:
              type: string
              enum: [dangling]
            removed:
              type: array
              items:
                type: string
            blobsRemoved:
              type: integer
            staleCleaned:
              type: integer
...